requirements.txt
beautifulsoup4==4.12.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
fonttools==4.39.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector-python==8.0.33
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
protobuf==3.20.3
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
simplejson==3.19.1
six==1.16.0
soupsieve==2.4
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
